server/requirements.txt
fastapi
uvicorn[standard]

langchain
langchain_community
langchain_core
langchain_groq

chromadb

sentence-transformers

PyMuPDF 

python-dotenv

pydantic
requests

loguru
aiofiles

python-multipart
deep-translator
langdetect
youtube-transcript-api
pymongo[srv]


# -U "psycopg[binary,pool]" langgraph langgraph-checkpoint-postgres